feat: Refactor SlackNotifier integration and remove Capistrano integration
